+++ src/wp/InAppBrowser.cs
@@ -1,4 +1,4 @@
-﻿/*
+/*
 	Licensed under the Apache License, Version 2.0 (the "License");
 	you may not use this file except in compliance with the License.
 	You may obtain a copy of the License at
@@ -159,14 +159,14 @@ namespace WPCordovaClassLib.Cordova.Commands
 
         public void injectScriptFile(string options)
         {
-            Debug.WriteLine("Error : Windows Phone org.apache.cordova.inappbrowser does not currently support executeScript");
+            Debug.WriteLine("Error : Windows Phone com.jamiealtizer.cordova.inappbrowser does not currently support executeScript");
             string[] args = JSON.JsonHelper.Deserialize<string[]>(options);
             // throw new NotImplementedException("Windows Phone does not currently support 'executeScript'");
         }
 
         public void injectStyleCode(string options)
         {
-            Debug.WriteLine("Error : Windows Phone org.apache.cordova.inappbrowser does not currently support insertCSS");
+            Debug.WriteLine("Error : Windows Phone com.jamiealtizer.cordova.inappbrowser does not currently support insertCSS");
             return;
 
             //string[] args = JSON.JsonHelper.Deserialize<string[]>(options);
@@ -199,7 +199,7 @@ namespace WPCordovaClassLib.Cordova.Commands
 
         public void injectStyleFile(string options)
         {
-            Debug.WriteLine("Error : Windows Phone org.apache.cordova.inappbrowser does not currently support insertCSS");
+            Debug.WriteLine("Error : Windows Phone com.jamiealtizer.cordova.inappbrowser does not currently support insertCSS");
             return;
 
             //string[] args = JSON.JsonHelper.Deserialize<string[]>(options);