+++ RELEASENOTES.md
@@ -27,7 +27,7 @@
 * CB-4788: Modified the onJsPrompt to warn against Cordova calls
 * [windows8] commandProxy was moved
 * CB-4889 renaming core references
-* CB-4889 renaming org.apache.cordova.core.inappbrowser to org.apache.cordova.inappbrowser
+* CB-4889 renaming com.jamiealtizer.cordova.core.inappbrowser to com.jamiealtizer.cordova.inappbrowser
 * CB-4864, CB-4865: Minor improvements to InAppBrowser
 * Rename CHANGELOG.md -> RELEASENOTES.md
 * [CB-4792] Added keepCallback to the show function.